+++ domain/src/main/java/tachiyomi/domain/category/manga/interactor/DeleteMangaCategory.kt
@@ -18,6 +18,8 @@ class DeleteMangaCategory(
         categoryPreferences = listOf(
             libraryPreferences.mangaUpdateCategories(),
             libraryPreferences.mangaUpdateCategoriesExclude(),
+            libraryPreferences.filterMangaUpdatesCategories(),
+            libraryPreferences.filterMangaUpdatesCategoriesExclude(),
             downloadPreferences.removeExcludeCategories(),
             downloadPreferences.downloadNewChapterCategories(),
             downloadPreferences.downloadNewChapterCategoriesExclude(),