+++ domain/src/main/java/tachiyomi/domain/updates/anime/repository/AnimeUpdatesRepository.kt
@@ -7,7 +7,12 @@ interface AnimeUpdatesRepository {
 
     suspend fun awaitWithSeen(seen: Boolean, after: Long, limit: Long): List<AnimeUpdatesWithRelations>
 
-    fun subscribeAllAnimeUpdates(after: Long, limit: Long): Flow<List<AnimeUpdatesWithRelations>>
+    fun subscribeAllAnimeUpdatesWithCategoryFilter(
+        after: Long,
+        limit: Long,
+        includedCategories: List<Long>,
+        excludedCategories: List<Long>,
+    ): Flow<List<AnimeUpdatesWithRelations>>
 
     fun subscribeWithSeen(seen: Boolean, after: Long, limit: Long): Flow<List<AnimeUpdatesWithRelations>>
 }