+++ domain/src/main/java/tachiyomi/domain/updates/manga/interactor/GetMangaUpdates.kt
@@ -13,8 +13,17 @@ class GetMangaUpdates(
         return repository.awaitWithRead(read, after, limit = 500)
     }
 
-    fun subscribe(instant: Instant): Flow<List<MangaUpdatesWithRelations>> {
-        return repository.subscribeAllMangaUpdates(instant.toEpochMilli(), limit = 500)
+    fun subscribe(
+        instant: Instant,
+        includedCategories: List<Long>,
+        excludedCategories: List<Long>,
+    ): Flow<List<MangaUpdatesWithRelations>> {
+        return repository.subscribeAllMangaUpdatesWithCategoryFilter(
+            after = instant.toEpochMilli(),
+            limit = 500,
+            includedCategories = includedCategories,
+            excludedCategories = excludedCategories,
+        )
     }
 
     fun subscribe(read: Boolean, after: Long): Flow<List<MangaUpdatesWithRelations>> {