+++ domain/src/main/java/tachiyomi/domain/updates/manga/repository/MangaUpdatesRepository.kt
@@ -7,7 +7,12 @@ interface MangaUpdatesRepository {
 
     suspend fun awaitWithRead(read: Boolean, after: Long, limit: Long): List<MangaUpdatesWithRelations>
 
-    fun subscribeAllMangaUpdates(after: Long, limit: Long): Flow<List<MangaUpdatesWithRelations>>
+    fun subscribeAllMangaUpdatesWithCategoryFilter(
+        after: Long,
+        limit: Long,
+        includedCategories: List<Long>,
+        excludedCategories: List<Long>,
+    ): Flow<List<MangaUpdatesWithRelations>>
 
     fun subscribeWithRead(read: Boolean, after: Long, limit: Long): Flow<List<MangaUpdatesWithRelations>>
 }